# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.CreateImageTweetItem:
    properties:
      image_url:
        example: https://example.com/image.jpg
        type: string
      user_id:
        example: 1
        type: integer
    type: object
  controllers.CreateImageTweetResponse:
    properties:
      tweet:
        $ref: '#/definitions/controllers.CreateImageTweetItem'
    type: object
...
  controllers.CreateTweetInput:
    properties:
      content:
        example: Hello, world!
        type: string
    required:
    - content
    type: object
  controllers.CreateTweetItem:
    properties:
      content:
        example: Hello, world!
        type: string
      user_id:
        example: 1
        type: integer
    type: object
  controllers.CreateTweetResponse:
    properties:
      tweet:
        $ref: '#/definitions/controllers.CreateTweetItem'
    type: object
  controllers.CreateUserProfileBody:
    properties:
      bio:
        example: 自己紹介文です
        maxLength: 160
        type: string
      image_url:
        example: https://example.com/avatar.png
        type: string
      location:
        example: Tokyo
        maxLength: 10
        type: string
      name:
        example: たろう
        maxLength: 20
        type: string
    type: object
  controllers.ErrorResponse:
    properties:
      error:
        example: エラーメッセージ
        type: string
    type: object
...
  controllers.GetTweetResponse:
    properties:
      tweet:
        $ref: '#/definitions/controllers.SwaggerTweet'
    type: object
...
  controllers.GetUserTweetsResponse:
    properties:
      has_more:
        example: false
        type: boolean
      next_cursor:
        type: integer
      tweets:
        items:
          $ref: '#/definitions/controllers.SwaggerTweet'
        type: array
    type: object
...
  controllers.StatusOKResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
...
  controllers.SwaggerTweet:
    properties:
      content:
        example: Hello, world!
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      like_count:
        example: 0
        type: integer
      retweet_count:
        example: 0
        type: integer
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user_id:
        example: 1
        type: integer
    type: object
...
  controllers.SwaggerUserProfile:
    properties:
      bio:
        example: 自己紹介文のサンプル文です
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      image_url:
        example: https://example.com/avatar.png
        type: string
      location:
        example: 東京
        type: string
      name:
        example: Taro007
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user_id:
        example: 1
        type: integer
    type: object
  controllers.UpdateUserProfileBody:
    properties:
      bio:
        example: 自己紹介文です
        maxLength: 160
        type: string
      image_url:
        example: https://example.com/avatar.png
        type: string
      location:
        example: Tokyo
        maxLength: 10
        type: string
      name:
        example: たろう
        maxLength: 20
        type: string
    type: object
  controllers.UserProfileResponse:
    properties:
      profile:
        $ref: '#/definitions/controllers.SwaggerUserProfile'
    type: object
host: localhost:8080
info:
  contact: {}
  description: Twitter クローン API
  title: Golang Twitter API
  version: "1.0"
paths:
...
  /profile:
    post:
      consumes:
      - application/json
      description: ログイン中のユーザーのプロフィールを作成する
      parameters:
      - description: プロフィール情報
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateUserProfileBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.UserProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - SessionAuth: []
      summary: プロフィール作成
      tags:
      - profiles
    put:
      consumes:
      - application/json
      description: ログイン中のユーザーのプロフィールを更新する
      parameters:
      - description: プロフィール情報
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateUserProfileBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.UserProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - SessionAuth: []
      summary: プロフィール更新
      tags:
      - profiles
...
  /tweets:
    post:
      consumes:
      - application/json
      description: 認証済みユーザーとしてツイートを投稿する
      parameters:
      - description: ツイート内容
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateTweetInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.CreateTweetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - SessionAuth: []
      summary: ツイート投稿
      tags:
      - tweets
  /tweets-image:
    post:
      consumes:
      - multipart/form-data
      description: 認証済みユーザーとして画像ファイルを投稿する
      parameters:
      - description: 画像ファイル（JPEG/PNG・5MB以下）
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.CreateImageTweetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - SessionAuth: []
      summary: 画像投稿
      tags:
      - tweets
  /tweets/{id}:
    delete:
      description: 認証済みユーザーが自分のツイートを削除する（他人のツイートは削除できない）
      parameters:
      - description: ツイートID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.StatusOKResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - SessionAuth: []
      summary: ツイート削除
      tags:
      - tweets
    get:
      description: 指定IDのツイートを取得する
      parameters:
      - description: ツイートID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.GetTweetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: ツイート取得
      tags:
      - tweets
...
  /users/{user_id}/profile:
    get:
      description: 指定ユーザーのプロフィールを取得する
      parameters:
      - description: ユーザーID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.UserProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: プロフィール取得
      tags:
      - profiles
...
  /users/{user_id}/tweets:
    get:
      description: 指定ユーザーのツイートをカーソルページネーションで取得する
      parameters:
      - description: ユーザーID
        in: path
        name: user_id
        required: true
        type: integer
      - description: ページネーションカーソル（最後に取得したツイートID）
        in: query
        name: cursor
        type: integer
      - default: 20
        description: 取得件数（1〜100、デフォルト20）
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.GetUserTweetsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: ユーザーのツイート一覧取得
      tags:
      - tweets
securityDefinitions:
  SessionAuth:
    in: cookie
    name: golang_twitter_session
    type: apiKey
swagger: "2.0"